+++ moban/jinja2/tests/files.py
@@ -1,15 +1,12 @@
-from os.path import (
-    isabs,
-    isdir,
-    exists,
-    isfile,
-    islink,
-    ismount,
-    lexists,
-    samefile,
-)
+import sys
+from os.path import isabs, isdir, exists, isfile, islink, ismount, lexists
 
 from moban.jinja2.extensions import jinja_tests
+
+if sys.platform == "win32":
+    from moban.jinja2.tests.win32 import samefile
+else:
+    from os.path import samefile
 
 jinja_tests(
     is_dir=isdir,