+++ tests/test_copy_engine.py
@@ -12,7 +12,8 @@ class TestContentForwardEngine:
 
     def test_get_template(self):
         template_content = self.engine.get_template("copier-test01.csv")
-        eq_("test 01\n", template_content)
+        #  remove '\r' for windows
+        eq_("test 01\n", template_content.replace("\r", ""))
 
     def test_get_template_from_string(self):
         test_content = "simply forwarded"